update registration data added
