+++ silasdk/schema.py
@@ -332,6 +332,25 @@ Schema = [
         "postal_code": "",
         "country": ""
     }},
+    {"update_registration_data_msg": {
+        "header": {
+            "created": "",
+            "auth_handle": "",
+            "user_handle": ""
+        },
+        "uuid": "",
+        "email": "",
+        "phone": "",
+        "identity_alias": "",
+        "identity_value": "",
+        "address_alias": "",
+        "street_address_1": "",
+        "street_address_2": "",
+        "city": "",
+        "state": "",
+        "postal_code": "",
+        "country": ""
+    }},
     {"delete_registration_data_msg": {
         "header": {
             "created": "",
