+++ silasdk/users.py
@@ -211,6 +211,20 @@ class User():
             self, path, msg_type, payload, user_private_key)
         return response
 
+    def updateRegistrationData(self, registrationField, payload, user_private_key):
+        """
+        Args:
+            payload: registration data
+            user_private_key
+        Returns:
+            dict: response body (entity information)
+        """
+        path = endPoints["updateRegistrationData"] + registrationField
+        msg_type = "update_registration_data_msg"
+        response = message.postRequest(
+            self, path, msg_type, payload, user_private_key)
+        return response
+
     def deleteRegistrationData(self, registrationField, payload, user_private_key):
         """
         Args:
